[Nexters/covid-letter-web#170] ♻️ useProfileError 로 프로필 에러 로직 분리
